add update category (cont.)

+++ app/src/main/java/com/app/e_commerce_app/activities/MainActivity.kt
@@ -1,23 +1,17 @@
 package com.app.e_commerce_app.activities
 
 import android.os.Bundle
-import android.util.Log
 import android.view.View
-import androidx.activity.viewModels
 import androidx.databinding.DataBindingUtil
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupWithNavController
-import com.app.e_commerce_app.MyApplication
 import com.app.e_commerce_app.R
 import com.app.e_commerce_app.base.BaseActivity
-import com.app.e_commerce_app.base.BaseFragment
-import com.app.e_commerce_app.common.AppSharePreference
 import com.app.e_commerce_app.data.repository.TokenRepository
 import com.app.e_commerce_app.databinding.ActivityMainBinding
 import com.app.e_commerce_app.utils.Utils
-import com.app.e_commerce_app.viewmodel.UserViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
@@ -27,6 +21,7 @@ class MainActivity : BaseActivity() {
     private var _binding: ActivityMainBinding? = null
     private val binding get() = _binding!!
     private lateinit var controller: NavController
+
     @Inject
     lateinit var tokenRepository: TokenRepository
 
@@ -73,13 +68,6 @@ class MainActivity : BaseActivity() {
     override fun onStop() {
         super.onStop()
         val remember = tokenRepository.getRemember()
-        if(remember == null)
-            Log.d("CHECKMAIN", "BEFORE null")
-        else
-            Log.d("CHECKMAIN", "BEFORE ${remember.toString()}")
-
-
-        if (tokenRepository.getRemember() == false)
-            tokenRepository.removeToken()
+        if (!remember) tokenRepository.removeToken()
     }
 }
